requirements.txt
############################################################
# Core runtime
############################################################
# HTTP + utils
requests>=2.32.3
tqdm>=4.66.4

# Scraper stack
selenium>=4.24.0
websocket-client>=1.8.0

# Math/tokenization
sympy>=1.12
tiktoken>=0.7.0

############################################################
# LLM provider SDKs
############################################################
openai>=1.45.0
together>=1.3.0

############################################################
# Web app
############################################################
Flask>=3.0.0
python-dotenv>=1.0.1

############################################################
# Dev / Test
############################################################
pytest>=8.3.0
black>=24.8.0
ruff>=0.6.0

############################################################
# Visualization
############################################################
matplotlib>=3.9.0        # graphs in benchmarking tool and finalize_run_outputs

############################################################
# Optional extras (uncomment if needed)
############################################################
# anthropic>=0.36.0        # Anthropic adapter in benchmarking harness
# xai-sdk>=0.2.0           # xAI Grok adapter (preferred over HTTP fallback)
